unknown changes from android-tegra-nv-3.4
Rebase-Id: R940fad74c7e91ef3d1d3d589a48064ccb7335541

+++ arch/arm/mach-tegra/io.c
@@ -76,13 +76,13 @@ static struct map_desc tegra_io_desc[] __initdata = {
 		.type = MT_DEVICE,
 	},
 	{
-		.virtual = (unsigned long)IO_PPCS_VIRT,
+		.virtual = IO_PPCS_VIRT,
 		.pfn = __phys_to_pfn(IO_PPCS_PHYS),
 		.length = IO_PPCS_SIZE,
 		.type = MT_DEVICE,
 	},
 	{
-		.virtual = (unsigned long)IO_PCIE_VIRT,
+		.virtual = IO_PCIE_VIRT,
 		.pfn = __phys_to_pfn(IO_PCIE_PHYS),
 		.length = IO_PCIE_SIZE,
 		.type = MT_DEVICE,
